agrego busqueda y eliminacion de eventos

+++ src/main/java/unq/tpi/desapp/model/Product.java
@@ -8,7 +8,7 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    Integer id;
+    Long id;
 
     Double price;
     Integer amountLimit;
@@ -30,7 +30,7 @@ public class Product {
         this.amountLimit = amountLimit;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
